feat：【IoT 物联网】增加 OTA 升级任务的实现

+++ yudao-module-iot/yudao-module-iot-biz/src/main/java/cn/iocoder/yudao/module/iot/dal/dataobject/ota/IotOtaTaskDO.java
@@ -61,7 +61,7 @@ public class IotOtaTaskDO extends BaseDO {
     /**
      * 设备总数数量
      */
-    private Long deviceTotalCount;
+    private Integer deviceTotalCount;
     /**
      * 设备成功数量
      */
